"remove Converter.kt and add test-classes "
