+++ src/test/kotlin/datasource/CSVDataSourceTest.kt
@@ -1,0 +1,12 @@
+package datasource
+
+import org.junit.jupiter.api.Test
+
+import org.junit.jupiter.api.Assertions.*
+
+internal class CSVDataSourceTest {
+
+    @Test
+    fun getAllApps() {
+    }
+}
