+++ src/test/kotlin/datasource/CSVReaderTest.kt
@@ -1,0 +1,12 @@
+package datasource
+
+import org.junit.jupiter.api.Test
+
+import org.junit.jupiter.api.Assertions.*
+
+internal class CSVReaderTest {
+
+    @Test
+    fun getTableRows() {
+    }
+}
